model + modeldto done,   method create in backend + test done
