+++ backend/src/main/java/com/example/demo/todo/Model/Todo.java
@@ -8,16 +8,17 @@ import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document (collection = "tasks")
 @Data
-@NoArgsConstructor
 @AllArgsConstructor
 public class Todo {
 
     @Id
     private String id;
-    private String task = " ";
-    private String description = " ";
-    private TodoCategory category = TodoCategory.Me ;
-    private TodoStatus status = TodoStatus.Open;
+    private String task;
+    private String description;
+    private TodoCategory category;
+    private TodoStatus status;
     private String user;
 
 }
+
+
