+++ backend/src/main/java/com/example/demo/todo/ModelDTO/TodoDTO.java
@@ -1,0 +1,18 @@
+package com.example.demo.todo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class TodoDTO {
+
+    private String id;
+    private String subject;
+    private String description;
+    private String user;
+
+}
