--- backend/src/main/java/com/example/demo/todo/TodoDTO.java
@@ -1,9 +1,0 @@
-package com.example.demo.todo;
-
-public class TodoDTO {
-
-    private String subject;
-    private String description;
-    private String user;
-
-}
